util/remove/requirements.txt
opencv-python>=4.8.1
numpy>=1.26.0
Pillow>=10.0.0 